requirements.txt
textual>=0.1.18
pyyaml>=6.0
rapidfuzz>=1.10.0